requirements.txt
anyio==3.5.0
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
entrypoints==0.4
executing==0.8.3
fastapi==0.77.1
fonttools==4.33.3
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
#ymatplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
#opencv-python==4.5.5.64
orjson==3.6.8
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scipy==1.8.0
#seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
stack-data==0.2.0
starlette==0.19.1
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typing_extensions==4.2.0
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
